Detele database if exist

+++ 0x0D-SQL_introduction/2-remove_database.sql
@@ -1,3 +1,2 @@
 -- Deletes the database hbtn_0c_0
-IF EXISTS (SELECT name FROM master.dbo.sysdatabases WHERE name = N’hbtn_0c_0’) THEN
-DROP DATABASE `hbtn_0c_0`;
+IF EXISTS (SELECT name FROM master.dbo.sysdatabases WHERE name = ’hbtn_0c_0’) DROP DATABASE `hbtn_0c_0`;
